added sleep in script to wait for peers and chaincodes. renamed some files. added sample app into docker img

+++ delete-k8-resources.sh
@@ -1,6 +1,7 @@
 kubectl delete deployments $(kubectl get deployments --template '{{range .items}}{{.metadata.name}}{{"\n"}}{{end}}') &&
-kubectl delete services ca-imfreemobile-com couchdb orderer-imfreemobile-com peer0-org1-imfreemobile-com docker &&
-kubectl delete pv shared-pv &&
-kubectl delete pvc shared-pvc &&
-kubectl delete pvc docker-pvc &&
-kubectl delete pv docker-pv
+kubectl delete services ca-imfreemobile-com couchdb orderer-imfreemobile-com peer0-org1-imfreemobile-com &&
+
+if [ "$1" = "--include-volumes" ]; then
+    kubectl delete pv shared-pv &&
+    kubectl delete pvc shared-pvc
+fi
